refactor tests and rename schema file
